app/api/gateway_use/requirements.txt
alembic==0.9.1
APScheduler==3.5.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
configparser==3.5.0
Flask==1.0.2
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
googlemaps==2.5.1
gpio==0.2.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.6
MarkupSafe==1.0
mysql-connector==2.1.6
mysqlclient==1.3.12
netifaces==0.10.6
paho-mqtt==1.3.1
pycparser==2.18
pycryptodome==3.4.6
pycryptodomex==3.4.7
PyMySQL==0.7.11
pyserial==3.4
python-apt==1.1.0b1+ubuntu0.16.4.2
python-editor==1.0.3
python-magic==0.4.22
pytz==2017.3
redis==3.3.11
requests==2.18.4
rq==1.1.0
six==1.11.0
sort==0.1.0
spidev==3.2
SQLAlchemy==1.1.9
ssh-import-id==5.5
tzlocal==1.5.1
urllib3==1.22
Werkzeug==0.14.1
